Text Input Created (not complete)
I created a input box using the text area element, i then layed this
element over some premade transperent texts which was sent by the
practise view. The user can then type over the transperent text, using
it as a guidline. I also created a javescript function that disables the
enter key as this was allowing ther user to missalighn the cursor.