Difficulty Changing through fetch request.
In my practise.html file i used the fetch API in my javascript to send a
request to the Update_word_bank view, while also sending the
difficultyselected that the selecter drop down was currently on. This view gets the word assosiated with the difficulty and then sends this to a partial template called partials/wordbank.html. A json response is then sent back to the practise.html template rendering the partial template in the practise.html template. This allows me to change the words displayed on the practise.html template without reloading the page. I also changed some of the styling as the typed words werent lining up with the transparent words.