Results screen created
i have added a partial template that covers the screen when the
user types the certain mounts of characters. I did this with a fetch
request to the Results veiw and then sends back the loaded partial
template as the response. This is then used in the javascript for the  practise.html template.
I also added a way for the user to play again when the results screen
apeared, through pressing Enter. I used an event listener for this.