Wordbank, Accuracy, and  updated topic wordbanks
I added a wordbank conataining every word that is displayed to the user
for tyipng. This wordbanck contains definitions or eac word. I did
this by manuely typing the defnintions, or rather copy and pasteing them
into the html code. For the accuracy in views i used the present TText
string aswell as the typedOverlay string and then compared each
character for each allowing me to see what character were correct and
therfore calculate the accuracy. Which is displayed in my results
partial template, with wpm. I also restyled the results partial
temnplate. I updated the prexisting wordbanks adding more words to the
topics, aswell as changing some of the original words for more
techinical terms, as i feel this gives the user more of a challange.